[nelson][board] Add platform bus node and post-init composite node spec
In the Nelson board driver, add a node that forwards the platform bus
protocol, and a composite node spec for the post-init driver to bind to.
The PLATFORM_DEV_INSTANCE_ID bind property is used to disambiguate
between the Nelson board driver and the platform bus node of the
post-init driver. Functionality will be moved from the board driver to
the post-init driver in subsequent changes.

Also add a device-enumeration-test case for the new post-init node.

Bug: b/304828063
Test: CQ
Change-Id: I9cba762ef0339e1a367c3453fa432f859b87be86
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/957173
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>
Fuchsia-Auto-Submit: Braden Kell <bradenkell@google.com>