[ffx][config] Can detect config domains in isolate
This makes it so that when you set up an isolated environment context,
if you give it a current working directory it will attempt to find a
configuration domain from that path as in the detect logic.

The reasoning for this is that generally, in the situations where we
currently advise using config domains, you still want project-specific
config to be available even if you don't want user config and other
similar stuff breaking in.

Currently, out of tree scripts that use isolation are passing in the
hidden `--env-root` argument to achieve this effect, and it doesn't make
sense to have them do so all the time.

Fixed: b/290946797
Change-Id: I578c493f79b96d1c498b79106dd08a33826f0a34
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959571
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>