[i2cimpl] Update Transact FIDL method
Update the Transact method for the fuchsia.hardware.i2cimpl FIDL
protocol so that it takes a `type` parameter instead of a `data` and
`is_read` parameter. The `type` parameter is a union that can be set to
either `read_size` denoting the amount of bytes to read or
`write_data` containing the bytes to be written.

`data` and `is_read` parameters are removed because `data` is not used
if `is_read` is set to true in the FIDL version of the i2cimpl protocol.
Unlike the banjo version, the FIDL version of the Transact method
returns data that is read instead of inserting it into `data`.

This also makes the i2cimpl Transfer FIDL method more similar to the i2c
Transact FIDL method.

Bug: b/301004275
Change-Id: I635431f16624d966672e5659e82c9605f9689f49
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/957852
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>