[PowerManager][Test] Add CPU thermal model test
This change migrates thermal_integration_tests to use RealmBuilder to
launch the PowerManager component in the integration test.
The migrated test covers the CPU thermal model administered by
CpuControlHandler. For test_average_temperature, compared to the
original test, the number of warm up and test iterations are reduced,
and the tolerance for the final assertion is relaxed. This is to reduce
the duration of the test.

FakeClock and NamedTimer are used to sync the time between production
code and test.

Bug: b/295078019
Test: fx test //src/power/power-manager
Change-Id: I7a8ee2e308c4dcf0bb9d11c622a09c6928ac3a59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947899
Reviewed-by: Onath Dillinger <claridge@google.com>
Commit-Queue: Didi She <didis@google.com>