[bazel-sdk] - allow using HEAD for in-tree bazel builds
We want to be able to set HEAD for in-tree builds to match the way GN
builds but we don't want to be do this in OOT builds. This CL adds
a mechanism which lets our in-tree build override the sdk generator to
allow for HEAD to be used.

This will leak out to OOT users at the moment because they are just
using the in-tree SDK but in the future we will disable this with a GN
arg.

Change-Id: I5f5eb4d5b93334eef2de2316efac971aa89302bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/960999
Reviewed-by: Darren Chan <chandarren@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>