[audio][vim3] Fix type of errors returned
For the VIM3 audio driver, return DriverError (not
zx_status_t) for non-fatal errors in Composite methods.

Update SetElementState to return an error but not close
the protocol channel, if the specified element_id is
unrecognized.

Update SetTopology to succeed (if the one and only VIM3-
reported topology ID is passed) or fail with INVALID_ARGS
(if any other topology ID value is passed).

Test: fx test -ouv audio_driver_tests -- --admin

Change-Id: Iba1764336d99818686ee23fdcc4a54af3e1c4de1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961056
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>