[fx][analytics] Remove target collection
Stop reporting PRODUCT.BOARD. We may bring it back in the future when we
figure out the right way to handle it. Also the current argument parsing
in analytics is naive such that the first argument after fx set is
always treated as PRODUCT.BOARD and `--with packages` is reported while
`--with=packages` is not. We will fix those later.

Fixed: b/316215093
Change-Id: I2742d7a80b2ebde1d954647ceacf69692fc2a70c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961852
Commit-Queue: Rundong Du <rundongdu@google.com>
Reviewed-by: Bob Evans <bobevans@google.com>