[doc-checker] Add option for allowing fuchsia-src links
This adds an option to allow links to reference /fuchsia-src
which is usually not allowed. The preferred link would be to
have a file path to /docs/... which is mapped to fuchsia-src.

This flag bypasses that check and is intended to be used when running
doc-checker on a project that is not part of fuchisa.

Bug: b/313673560
Change-Id: I81172821255e1e0d17fc7a957dd3d3040060c772
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961854
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Reviewed-by: Kyo Lee <kyol@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>