[starnix] Make default mutable_state not require Arc<>
The mutable_state macro allows the method on the mutable part of a
struct to have access to the immutable part. Change the default type of
the immutable part to be a simple reference instead of an Arc.

Change-Id: I44dfa975aa0f673412c069defd4c11db46276182
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/962312
Reviewed-by: Fabio D'Urso <fdurso@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>