[magma] Add dep on LogSink shard
Magma isn't using test(), so can't automatically
include the shard that way through the new build rule.
This CL adds an explicit dep on the shard so that
this test continues to work when we remove
the build config which implicitly adds LogSink
to all executable targets in Fuchsia.

Bug: b/300157652
Change-Id: Iae49fcc0f980b539228f49b096ed6785b91a5ac2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961882
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>