[role_manager] Create the RoleManager
Creates the fuchsia.scheduler.RoleManager FIDL protocol along with the
associated service implementation. This will be used by both in tree and
OOT components to set profiles based on roles.

Note that this CL does not migrate users of the existing ProfileProvider
API to the RoleManager; that will be done in future CLs. Additionally,
it does not add support for the input/output role parameters, that will
also be added in future CLs.

Bug: 127718
Change-Id: Ic29213983f22ca58e14913a4960b69242a88ff1a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/954056
Reviewed-by: Laura Peskin <pesk@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
API-Review: Hunter Freyer <hjfreyer@google.com>