[build] Remove syslog hack
This removes one of the syslog hacks that
added the backend automatically as part of the build.
Targets now need to explicitly dep on :cpp instead.
After this and other hacks are removed, the concept of
backends can be removed from the build system.

Bug: b/300157652
Change-Id: Ibedcdd5394e42e3ddf24873eb960c6b0ffed6fc5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/962661
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>