[logs] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I33b2cd6686e8acd887cb6774e977600bed9c1bd7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/962819
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: David Fang <fangism@google.com>