[logs] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I48e22fdb202d4acfe038bce07395bf78dad2b742
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963114
Reviewed-by: Filip Filmar <fmil@google.com>
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: David Fang <fangism@google.com>