[power] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I84f712673df40691312b16366aa2be8aebe8fb92
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963117
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Didi She <didis@google.com>