[logs] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I8002990c6be89ade2b3a0605eaf4c9645bfdc1ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963202
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>