[camera] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I93ede2606aaf0ea85ba58ff516f2176b67fcba90
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963392
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>