[build_info] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I64aa0f505ae857686d016d724c2b2db54449401d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963209
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>