[connectivity] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I2b25d27f42f85a5b306e3d317a4c44efd93df314
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963280
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>