[component] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I86f572ab26ffd67418f653f7ab9a4330f28fa69e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963211
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>