[watchdog] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: Iac0061d7b73dec29a3c3c96f78c35441f1f4ef50
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963206
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>