[codec] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I1775b90a1f63b8529ebdd9d27a0f59678e2a89c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963203
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>