[fsl] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I6a8e8bcf590ad2c92deb2b3eacaa83e40bf2ade1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963207
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>