[bluetooth] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: Id3a4b84c80d91b1c3b70f5dfb35a2dfe3d3985c0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963281
Reviewed-by: Ali Saeed <saeedali@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Garratt Gallagher <garratt@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>