[driver_test_realm] Wait for /dev/sys/test
The driver test realm binds a /dev/sys/test device during setup, but
this can be racy with the later teardown of the test realm.  If the test
completes quickly and the driver_manager starts to suspend, then the
later call to add the device will fail and emit an error log (which
fails unit tests).

Fixed: 316579125
Multiply: devfs-test.cm
Change-Id: I0de06101a3d10a7e2d919052379723c34477c724
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963955
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>