[ui] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: Ic3cdc01a2f31cb0f26179997b53e60037365a42b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964254
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Yilong Li <liyl@google.com>