[performance] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: Id157428a649a1cbcabb09da5b04cdf16027ec84e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964082
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>