[developer] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Bug: b/300157652
Change-Id: I2d512f49cc0f43f9327b794789438820509b5cb0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964088
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>