[devicetree] Use node name as pdev name
Pdev composed name using vid:pid:did, but this doesn't help with
devicetree case as vid:pid:did are not set. It hinders writing
enumeration tests that match the devicetree and also makes it harder to
debug devicetree node bindings at runtime. This CL changes pdev to use
the node name passed as it is for devicetree nodes. Ideally pdev can be
updated to do this for all cases, but the dependency on the path is
extensive and changes would have to be done in stages. For now, updating
the logic for devicetrees alone.

Test: vim3-devivetree-test
Change-Id: Id1f2e1c02be66df9bbdb7a3a628ecc2f0713c864
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964081
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>