[sandbox] Ability to fully drain a Receiver
Receiver today holds a sender referencing itself. While this is
convenient in that one could always get a sender from the receiver, this
takes away the ability to know if all external sender references are
dropped. We won't be able to get an EOF from a receiver because the
receiver always assumes another sender is referencing it.

This CL moves that internal sender reference out such that Receiver::new
always creates a pair of sender and receiver. When the last sender is
dropped, Receive::receive will return None.

This is a performance optimization to avoid monitoring an indefinitely
increasing list of Receivers in fxrev.dev/963411.

Change-Id: I8f9716f0ae345be5526b920046de727f590b969e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964079
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>