[starnix] Allow thread creation not to wait on pre-run
Following the change of CurrentTask creation so that it is done on the
final thread running the task, `execute_task` has been changed to
execute a `pre_run` function on the thread and using a channel to return
a result.

This is not needed for the clone operation but only for tests and for
container setup

This CL change `execute_task` to not wait on the result of `pre_run` and
instead introduce `execute_task_with_prerun_result` for the case where
it is needed.

Fixed: b/315453253
Change-Id: Id20b14676e75db51025e0308b1925cead86a3c7b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964632
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>