[fxl] Add fxl_cli target
Drivers use fxl, and can't depend on syslog.
Due to this limitation, the portions of fxl
that depend on syslog have to be removed from the main
library. This is a breaking change, so requires a soft
transition. This CL adds a new fxl_cli target which deps
on fxl, to facilitate the migration of things using cli
parts of fxl to, at which point the cli functionality
can be removed from fxl and moved into fxl_cli.

Bug: b/300157652
Change-Id: I662f8046b47d8c9266ce7eda0048d3a83ff22916
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964972
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Adam Barth <abarth@google.com>