[netstack3] Update SendOptions for datagram socket
Update the datagram socket layer so that the underlying `SendOptions`
implementation is `SocketHopLimits` rather than `IpOptions`.

This paves a way for follow up CL which uses different `SocketHopLimits`
for dualstack sockets sending to the other stack. `IpOptions` is
insufficient because it must be parameterized by an `I` representing the
"Socket IP Version" (e.g. IPv6 in the dualstack case), while the
`SendOptions` must be parameterized by an `I` representing the "Wire IP
Version" (e.g. IPv4 in the dualstack case)

Bug: b/42182532

Change-Id: Id81b378907dd22bc47f8f65c9d222051a4dced8b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/965032
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>