[assembly] Implement configuration capabilities
With this change, assembly is responsible for creating
a package named 'config' and adding it to the boot package
list. This package contains a single CML that exposes
configuration values.

To prove this concept out, a single configuration value
is added for the driver index. More configuration values
will be converted in a followup CL.

Bug: b/317057378
Change-Id: Icb67ac6711334d8acf8416d72079e55ba56344ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/956056
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>