[netstack3] Rename bindings context types
Following SyncCtx => CoreCtx and NonSyncCtx => BindingsCtx.

Shorten name of `BindingsNonSyncCtxImpl` to `BindingsCtx`. The `Impl`
bit was not diambiguating it with anything else and it follows the
convention of `Context` is trait and `Ctx` is concrete type.

Bug: b/42083910
Change-Id: I85a4c51aeca48a18ac72f9f4efef360c59459169
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/965628
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>