[compat] Fix breakage with compat device release hook impl.
There was an earlier change to delay release until after the
dispatcher shutsdown, to match DFv1 behaviour:
https://fuchsia-review.git.corp.google.com/c/fuchsia/+/961879

That broke the upgrade test for
smart_display_eng.astro-release-e2e-isolated-omaha,
where a driver rebind gets stuck during paving.

To fix this, we try to match the previous passing behaviour
as much as possible - only delay the release op, but the
rest of the compat device object can destruct as normal.
The actual device's context is stored along with the release op
to be called later.

Bug: 317784515

Change-Id: Ib834fb117023e9d90ab45530907dba8405ce97e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/965633
Reviewed-by: Justin Mattson <jmatt@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>