[i2c] Don't use banjo version of i2cimpl
Make the i2c driver not use the banjo version of the i2cimpl protocol.
It originally had a fallback to connect to the fidl version of the
i2cimpl protocol if it failed to connect to the banjo version. This
fallback will be default behavior now.

Bug: b/301004275
Change-Id: Ib543635f95ea3c640c899b67f3162ca6f8e0ea27
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/956882
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>