[wlan][hw-sim] Add drivers_only topology for new config API
This begins adding the new configuration API for the WLAN test
realm starting with the drivers_only test realm topology.

This also adds unit tests to the hw-sim realm factory to show that the
realm factory requires all fields specified for the DriversOnly
topology.

test.wlan.realm is added to the "internal" category and explicit
versioning attributes are added to the API to denote which fields/types
are deprecated. This will also make it easier to move this to
partner_internal in the future.

Bug: b/302713436
Test: fx test realm-factory-test simulate-scan
Change-Id: I5af523e15edfab16cf42f906946d13570b6e165f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961033
Reviewed-by: Charles Celerier <chcl@google.com>
Commit-Queue: John Tan <johntan@google.com>